test: move utils funcs to conftest
